Move wide and deep model to ranking.py.
