+++ merlin/models/tf/__init__.py
@@ -102,7 +102,7 @@ from merlin.models.tf.metrics.topk import (
 )
 from merlin.models.tf.models import benchmark
 from merlin.models.tf.models.base import BaseModel, Model, RetrievalModel
-from merlin.models.tf.models.ranking import DCNModel, DeepFMModel, DLRMModel
+from merlin.models.tf.models.ranking import DCNModel, DeepFMModel, DLRMModel, WideAndDeepModel
 from merlin.models.tf.models.retrieval import (
     MatrixFactorizationModel,
     TwoTowerModel,
@@ -219,6 +219,7 @@ __all__ = [
     "DLRMModel",
     "DCNModel",
     "DeepFMModel",
+    "WideAndDeepModel",
     "losses",
     "LossType",
     "sample_batch",
