+++ merlin/models/tf/models/ranking.py
@@ -1,4 +1,5 @@
-from typing import List, Optional, Union
+import warnings
+from typing import Dict, List, Optional, Union
 
 import tensorflow as tf
 
@@ -8,7 +9,7 @@ from merlin.models.tf.blocks.interaction import FMPairwiseInteraction
 from merlin.models.tf.blocks.mlp import MLPBlock
 from merlin.models.tf.core.aggregation import ConcatFeatures, StackFeatures
 from merlin.models.tf.core.base import Block
-from merlin.models.tf.core.combinators import ParallelBlock
+from merlin.models.tf.core.combinators import ParallelBlock, TabularBlock
 from merlin.models.tf.core.transformations import CategoricalOneHot
 from merlin.models.tf.inputs.base import InputBlock
 from merlin.models.tf.inputs.continuous import ContinuousFeatures
@@ -90,7 +91,7 @@ def DCNModel(
     prediction_tasks: Optional[
         Union[PredictionTask, List[PredictionTask], ParallelPredictionBlock]
     ] = None,
-    **kwargs
+    **kwargs,
 ) -> Model:
     """Create a model using the architecture proposed in DCN V2: Improved Deep & Cross Network [1].
     See Eq. (1) for full-rank and Eq. (2) for low-rank version.
@@ -171,7 +172,7 @@ def DeepFMModel(
         Union[PredictionTask, List[PredictionTask], ParallelPredictionBlock]
     ] = None,
     embedding_option_kwargs: dict = {},
-    **kwargs
+    **kwargs,
 ) -> Model:
     """DeepFM-model architecture.
 
@@ -212,7 +213,7 @@ def DeepFMModel(
         embedding_options=EmbeddingOptions(
             embedding_dim_default=embedding_dim, **embedding_option_kwargs
         ),
-        **kwargs
+        **kwargs,
     )
 
     pairwise_block = FMPairwiseInteraction().prepare(aggregation=StackFeatures(axis=-1))
@@ -233,5 +234,153 @@ def DeepFMModel(
 
     prediction_tasks = parse_prediction_tasks(schema, prediction_tasks)
     model = Model(deep_fm, prediction_tasks)
+
+    return model
+
+
+def WideAndDeepModel(
+    schema: Schema,
+    deep_block: Block,
+    embedding_dims: Optional[Dict[str, int]] = None,
+    embedding_dim_default: Optional[int] = None,
+    wide_schema: Optional[Schema] = None,
+    deep_schema: Optional[Schema] = None,
+    wide_preprocess: Optional[Block] = None,
+    deep_input_block: Optional[Block] = None,
+    wide_input_block: Optional[Block] = None,
+    prediction_tasks: Optional[
+        Union[PredictionTask, List[PredictionTask], ParallelPredictionBlock]
+    ] = None,
+    embedding_option_kwargs: dict = {},
+    **kwargs,
+) -> Model:
+    """Wide-and-Deep-model architecture.
+
+    Example Usage::
+        wide_deep = ml.benchmark.WideAndDeepModel(
+                schema,
+                deep_block=ml.MLPBlock([32, 16]),
+                embedding_dims={"user_catetory": 32},
+                embedding_dim_default=64,
+                wide_schema=wide_schema,
+                prediction_tasks=ml.BinaryClassificationTask("click"),
+            )
+        wide_deep.compile(optimizer="adam")
+        wide_deep.fit(train_data, epochs=10)
+
+    References
+    ----------
+    [1] Cheng, Koc, Harmsen, Shaked, Chandra, Aradhye, Anderson et al. "Wide & deep learning for
+    recommender systems." In Proceedings of the 1st workshop on deep learning for recommender
+    systems, pp. 7-10. (2016).
+
+    Parameters
+    ----------
+    schema : Schema
+        The `Schema` with the input features
+    deep_block: Block
+        Block (structure) of deep model.
+    embedding_dims : Optional[dict]
+        Dimensions of the embeddings, in order to specify different dimensions for different
+        features, for features not specified by this dict, default dimension
+        (embedding_dim_default) would be used to set their embeddings.by default None, then
+        embedding_dim_default would be the fixed embedding dimension for all features
+    embedding_dim : Optional[int]
+        Default dimension of feateures not specified in embedding_dims
+    wide_schema : Optional[Schema]
+        The 'Schema' of input features for wide model, by default no features would be sent to
+        wide model, and the model would become a pure deep model, if specified, only features
+        in wide_schema would be sent to wide model
+    deep_schema : Optional[Schema]
+        The 'Schema' of input features for deep model, by default all features would be sent to
+        deep model. deep_schema and wide_schema could contain the same features
+    wide_preprocess : Optional[Block]
+        Transformation block for preprocess data in wide model. Such as CategoricalOneHot,
+        CategoryEncoding, HashedCross, and HashedCrossAll, please note the schema of transformation
+        block should be the same as the wide_schema.  By default None.
+        For example:
+            ```python
+            # CategoricalOneHot as preprocess for wide model
+            import merlin.models.tf as ml
+            model = ml.benchmark.WideAndDeepModel(
+                schema = schema,
+                wide_schema=wide_schema,
+                deep_schema=deep_schema,
+                wide_preprocess = ml.CategoricalOneHot(wide_schema)
+                deep_block=ml.MLPBlock([32, 16]),
+                prediction_tasks=ml.BinaryClassificationTask("click"),
+            )
+
+            # HashedCross as preprocess for wide model
+            model = ml.benchmark.WideAndDeepModel(
+                schema = schema,
+                wide_schema=wide_schema,
+                deep_schema=deep_schema,
+                wide_preprocess = ml.HashedCross(wide_schema, num_bins=1000),
+                deep_block=ml.MLPBlock([32, 16]),
+                prediction_tasks=ml.BinaryClassificationTask("click"),
+            )
+            ```
+    prediction_tasks: Optional[Union[PredictionTask, List[PredictionTask], ParallelPredictionBlock]
+        The prediction tasks to be used, by default this will be inferred from the Schema.
+    embedding_option_kwargs: Optional[dict]
+        Additional arguments to provide to `EmbeddingOptions` object for embeddings tables setting.
+        Defaults to {}
+
+    Returns
+    -------
+    Model
+
+    """
+
+    prediction_tasks = parse_prediction_tasks(schema, prediction_tasks)
+
+    if embedding_dims and embedding_dim_default:
+        embedding_options = EmbeddingOptions(
+            embedding_dims=embedding_dims,
+            embedding_dim_default=embedding_dim_default,
+            **embedding_option_kwargs,
+        )
+    elif embedding_dims:
+        embedding_options = EmbeddingOptions(
+            embedding_dims=embedding_dims, **embedding_option_kwargs
+        )
+    elif embedding_dim_default:
+        embedding_options = EmbeddingOptions(
+            embedding_dim_default=embedding_dim_default, **embedding_option_kwargs
+        )
+    else:
+        embedding_options = EmbeddingOptions(**embedding_option_kwargs)
+
+    if not wide_schema:
+        warnings.warn("If not specify wide_schema, NO feature would be sent to wide model")
+        wide_schema = None
+
+    if not deep_schema:
+        deep_schema = schema
+
+    if not deep_input_block:
+        if len(deep_schema) > 0:
+            deep_input_block = InputBlock(
+                deep_schema,
+                embedding_options=embedding_options,
+                **kwargs,
+            )
+    deep_body = deep_input_block.connect(deep_block).connect(
+        MLPBlock([1], no_activation_last_layer=True)
+    )
+
+    if not wide_input_block:
+        if len(wide_schema) > 0:
+            wide_input_block = ParallelBlock(
+                TabularBlock.from_schema(schema=wide_schema, pre=wide_preprocess),
+                is_input=True,
+                aggregation="concat",
+            )
+    wide_body = wide_input_block.connect(MLPBlock([1], no_activation_last_layer=True))
+
+    branches = {"wide": wide_body, "deep": deep_body}
+    wide_and_deep_body = ParallelBlock(branches, aggregation="element-wise-sum")
+    model = Model(wide_and_deep_body, prediction_tasks)
 
     return model
