+++ tests/unit/tf/models/test_benchmark.py
@@ -17,7 +17,6 @@
 import pytest
 
 import merlin.models.tf as ml
-from merlin.dag.selector import ColumnSelector
 from merlin.models.tf.utils import testing_utils
 
 
@@ -32,42 +31,3 @@ def test_ncf_model(ecommerce_data, run_eagerly):
     )
 
     testing_utils.model_test(model, ecommerce_data, run_eagerly=run_eagerly)
-
-
-@pytest.mark.parametrize("run_eagerly", [True, False])
-def test_wide_deep_model(music_streaming_data, run_eagerly):
-
-    # prepare wide_schema
-    selector = ColumnSelector(["country"])
-    wide_schema = music_streaming_data.schema.select(selector)
-    selector = ColumnSelector(["country", "user_age"])
-    deep_schema = music_streaming_data.schema.select(selector)
-
-    model = ml.benchmark.WideAndDeepModel(
-        music_streaming_data.schema,
-        embedding_dim_default=64,
-        wide_schema=wide_schema,
-        deep_schema=deep_schema,
-        deep_block=ml.MLPBlock([32, 16]),
-        prediction_tasks=ml.BinaryClassificationTask("click"),
-    )
-
-    testing_utils.model_test(model, music_streaming_data, run_eagerly=run_eagerly)
-
-
-@pytest.mark.parametrize("run_eagerly", [True, False])
-def test_wide_deep_embedding_dims_dict(music_streaming_data, run_eagerly):
-
-    # prepare wide_schema
-    selector = ColumnSelector(["country"])
-    wide_schema = music_streaming_data.schema.select(selector)
-
-    model = ml.benchmark.WideAndDeepModel(
-        music_streaming_data.schema,
-        embedding_dims={"country": 32, "user_genres": 128},
-        wide_schema=wide_schema,
-        deep_block=ml.MLPBlock([32, 16]),
-        prediction_tasks=ml.BinaryClassificationTask("click"),
-    )
-
-    testing_utils.model_test(model, music_streaming_data, run_eagerly=run_eagerly)
