+++ tests/unit/tf/models/test_ranking.py
@@ -169,3 +169,75 @@ def test_serialization_model(ecommerce_data: Dataset, prediction_task):
     )
 
     testing_utils.model_test(model, ecommerce_data, reload_model=True)
+
+
+@pytest.mark.parametrize("run_eagerly", [True, False])
+def test_wide_deep_model(music_streaming_data, run_eagerly):
+
+    # prepare wide_schema
+    wide_schema = music_streaming_data.schema.select_by_name(["country"])
+    deep_schema = music_streaming_data.schema.select_by_name(["country", "user_age"])
+
+    model = ml.WideAndDeepModel(
+        music_streaming_data.schema,
+        embedding_dim_default=64,
+        wide_schema=wide_schema,
+        deep_schema=deep_schema,
+        deep_block=ml.MLPBlock([32, 16]),
+        prediction_tasks=ml.BinaryClassificationTask("click"),
+    )
+
+    testing_utils.model_test(model, music_streaming_data, run_eagerly=run_eagerly)
+
+
+@pytest.mark.parametrize("run_eagerly", [True, False])
+def test_wide_deep_model_categorical_one_hot(ecommerce_data, run_eagerly):
+
+    wide_schema = ecommerce_data.schema.select_by_name(names=["user_categories", "item_category"])
+    deep_schema = ecommerce_data.schema
+
+    model = ml.WideAndDeepModel(
+        ecommerce_data.schema,
+        wide_schema=wide_schema,
+        deep_schema=deep_schema,
+        wide_preprocess=ml.CategoricalOneHot(wide_schema),
+        deep_block=ml.MLPBlock([32, 16]),
+        prediction_tasks=ml.BinaryClassificationTask("click"),
+    )
+
+    testing_utils.model_test(model, ecommerce_data, run_eagerly=run_eagerly)
+
+
+@pytest.mark.parametrize("run_eagerly", [True, False])
+def test_wide_deep_model_hashed_cross(ecommerce_data, run_eagerly):
+
+    wide_schema = ecommerce_data.schema.select_by_name(names=["user_categories", "item_category"])
+    deep_schema = ecommerce_data.schema
+
+    model = ml.WideAndDeepModel(
+        ecommerce_data.schema,
+        wide_schema=wide_schema,
+        deep_schema=deep_schema,
+        wide_preprocess=ml.HashedCross(wide_schema, 1000),
+        deep_block=ml.MLPBlock([32, 16]),
+        prediction_tasks=ml.BinaryClassificationTask("click"),
+    )
+
+    testing_utils.model_test(model, ecommerce_data, run_eagerly=run_eagerly)
+
+
+@pytest.mark.parametrize("run_eagerly", [True, False])
+def test_wide_deep_embedding_dims_dict(music_streaming_data, run_eagerly):
+
+    # prepare wide_schema
+    wide_schema = music_streaming_data.schema.select_by_name(["country"])
+
+    model = ml.WideAndDeepModel(
+        music_streaming_data.schema,
+        embedding_dims={"country": 32, "user_genres": 128},
+        wide_schema=wide_schema,
+        deep_block=ml.MLPBlock([32, 16]),
+        prediction_tasks=ml.BinaryClassificationTask("click"),
+    )
+
+    testing_utils.model_test(model, music_streaming_data, run_eagerly=run_eagerly)
